==> 10_regression/yield-tables-regression/notebooks/00_initial_exploration.ipynb <==
# Initial Exploration

This notebook includes initial exploration of the dataset, for a structured data science project, see the following notebooks. If you want to get an initial idea of the dataset, check this notebook.

import pandas as pd

# Load the dataset
df = pd.read_csv("../data/raw/yield_tables.csv", sep=";")

print(df.info())

df.describe().T

df.isna().sum()

## Summary

The dataset is consistent but includes many null values. But after dropping all the null values, the size of the dataset should still be sufficient to train a regression model.

The columns are all `float` or `int` types, except for `mean_annual_increment` which is an `object` type, which is weird - when looking at the dataset it seems to be a number - anyway, we will not use it in our regression model since we are not interested in the growth of trees over time).
