==> 10_regression/yield-tables-regression/notebooks/04_leakage_check.ipynb <==
# Detecting leakage

It's essential to ensure that our model is not inadvertently using information from the future to make predictions about the past. This is known as "data leakage" and can lead to overly optimistic performance estimates during model evaluation.

import pandas as pd

# Load your prepared data from M3L1
df = pd.read_csv("../data/interim/cleaned_data.csv")
target_variable = "volume_per_ha"

print(df.head())

## Calculate Pearson Correlation Coefficient

If the Pearson correlation coefficient is higher than 0.95 we are facing potential data leakage. As the highest correlation is between `dbh` and the target variable (0.775), we don't need to take any action.

for col in df.columns:
    if col != "Area" and col != "Year":
        print(
            f"Correlation between {col} and {target_variable}: {df[[col, target_variable]].corr().iloc[0,1]}"
        )
